Move validation functions into separate module
