+++ euphonic/debye_waller.py
@@ -2,10 +2,10 @@ import inspect
 
 from pint import Quantity
 
-from euphonic import ureg, Crystal
-from euphonic.util import _check_constructor_inputs, _check_unit_conversion
+from euphonic.validate import _check_constructor_inputs, _check_unit_conversion
 from euphonic.io import (_obj_to_json_file, _obj_from_json_file,
                          _obj_to_dict, _process_dict)
+from euphonic import ureg, Crystal
 
 
 class DebyeWaller(object):
