+++ euphonic/force_constants.py
@@ -8,16 +8,15 @@ import scipy
 from scipy.linalg.lapack import zheev
 from scipy.special import erfc
 
-from euphonic import ureg
-from euphonic.crystal import Crystal
-from euphonic.qpoint_phonon_modes import QpointPhononModes
-from euphonic.util import (is_gamma, get_all_origins,
-                           _check_constructor_inputs,
-                           _check_unit_conversion,
-                           _get_supercell_relative_idx)
+from euphonic.validate import (
+    _check_constructor_inputs, _check_unit_conversion,
+    _ensure_contiguous_args, _ensure_contiguous_attrs)
 from euphonic.io import (_obj_to_json_file, _obj_from_json_file,
                          _obj_to_dict, _process_dict)
 from euphonic.readers import castep, phonopy
+from euphonic.util import (is_gamma, get_all_origins,
+                           _get_supercell_relative_idx)
+from euphonic import ureg, Crystal, QpointPhononModes
 
 
 class ImportCError(Exception):
@@ -385,8 +384,6 @@ class ForceConstants(object):
             if use_c:
                 try:
                     import euphonic._euphonic as euphonic_c
-                    from euphonic.util import (_ensure_contiguous_args,
-                                               _ensure_contiguous_attrs)
                 except ImportError:
                     warnings.warn((
                         'use_c=True is set, but the Euphonic\'s C '
